main-image/requirements.txt
alembic==1.3.3
amqp==2.5.2
apispec==1.3.3
argcomplete==1.11.1
attrs==19.3.0
Babel==2.8.0
blinker==1.4
cached-property==1.5.1
cachetools==4.0.0
cattrs==0.9.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.31
cycler==0.10.0
Cython==0.29.15
defusedxml==0.6.0
dill==0.3.1.1
dnspython==1.16.0
docutils==0.16
eventlet==0.25.1
Flask==1.1.1
Flask-Admin==1.5.4
Flask-AppBuilder==2.2.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-OAuthlib==0.9.5
Flask-OpenID==1.2.5
Flask-Script==2.0.6
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.16.0
gevent==1.4.0
gevent-websocket==0.10.1
graphviz==0.13.2
greenlet==0.4.15
grpcio==1.26.0
gunicorn==19.10.0
idna==2.8
importlib-metadata==1.5.0
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
json-merge-patch==0.2
jsonschema==3.2.0
kafka-python==2.0.0
kiwisolver==1.1.0
kombu==4.6.7
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.1
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.22.2
matplotlib==3.1.3
mongo-datatables==0.3.0
monotonic==1.5
numpy==1.18.1
oauthlib==2.1.0
pandas==0.25.3
pendulum==1.4.4
prison==0.1.2
protobuf==3.11.2
psutil==5.7.0
ptvsd==4.3.2
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
PyHDFS==0.3.1
PyJWT==1.7.1
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.6
pyrsistent==0.15.7
pyspark==2.4.5
python-daemon==2.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.11.2
python-socketio==4.4.0
python3-openid==3.1.0
pytz==2019.3
pytzdata==2019.3
PyYAML==5.3
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
setproctitle==1.1.10
simplejson==3.17.0
six==1.14.0
SQLAlchemy==1.3.13
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.1
tabulate==0.8.6
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
thrift==0.13.0
twilio==6.35.3
typing-extensions==3.7.4.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.8
vine==1.3.0
Werkzeug==0.16.0
WTForms==2.2.1
zipp==2.2.0
zope.deprecation==4.4.0
google-cloud-storage==1.26.0
mongo-datatables==0.3.0
gcsfs==0.6.0
